share code between android/ios
- references:
    - https://differential.com/insights/sharing-code-between-android-and-ios-in-react-native/
